osa4-d tehtävät

+++ controllers/blogRouter.js
@@ -1,47 +1,64 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 const express = require('express')
+const bcrypt = require('bcrypt')
 
 blogRouter.use(express.json())
 
 blogRouter.get('/api/blogs', async (request, response) => {
 
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
   response.status(200).json(blogs)  
 
 })
-  
-blogRouter.post('/api/blogs', async (request, response) => {
 
-  const body = request.body
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0
-  })
+blogRouter.post('/api/blogs', async (request, response, next) => {
+  try {
 
-  if (!blog.title || !blog.url) {
-    response.status(400).end()
-  } else {
-    const blogs = await blog.save()
-    response.status(201).json(blogs)
+    const body = request.body
+    const user = await User.findById(body.userId)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: request.token.id
+    })
+
+    if (!blog.title || !blog.url) {
+      next("missingInfo")
+    } else {
+      const blogs = await blog.save()
+      user.blogs = user.blogs.concat(blogs._id)
+      await user.save()
+      response.status(201).json(blogs)
+    }
+  } catch (error) {
+    next(error)
   }
-
+  
 })
 
 blogRouter.delete('/api/blogs/:id', async (request, response) => {
 
   try {
-    const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
+    const toBeDeleted = await Blog.findById(request.params.id)
 
-    if (deletedBlog) {
-      response.status(204).end()
+    if (toBeDeleted) {
+      if (toBeDeleted.user.toString() === request.token.id) {
+        const deleteBlog = await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+      } else {
+        response.status(404).end()
+      }    
     } else {
       response.status(404).end()
     }
     
   } catch (error) {
+    console.log(error);
     response.status(404).end()
   }
   
@@ -56,11 +73,9 @@ blogRouter.put('/api/blogs/:id', async (request, response) => {
       {new: true}
     )
 
-    console.log(blogs);
     response.status(201).json(blogs)
 
   } catch (error) {
-    console.log(error);
     response.status(404).end()
   }
 })
